midi preset (no cont. pedal)
The layout of the midi mapping is as close as i can get it to the original layout. I had to switch STACK and DIRECTION due to a limitation on the MIDI pedal i use. I might be able to work around that with a little work. This also does not have the expression pedal CC mapped to the output level.